Layout: vector graphics work update
VG: Removed cairo dependency completely, cleanup
VG: Implemented stroke and fill path rendering

Material: rewritten icons rendering (font-layout-like engine)
Stappler: fixed some memory leaks

Other: added generator sources for autogenerated code
- Color tables (SLColor.cpp)
- Icons (MaterialIconStorage.cpp)

+++ material/Material.h
@@ -55,7 +55,7 @@ class ResourceManager;
 class ResizeTo;
 class ResizeBy;
 
-enum class IconName : uint32_t;
+enum class IconName : uint16_t;
 
 class IconSprite;
 class IconProgress;
